# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "猫影短剧",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/config": {
      "post": {
        "tags": [
          "模型配置"
        ],
        "summary": "创建模型配置",
        "operationId": "create_config_api_config_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AiModelConfigCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseSchema_AiModelConfigOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "模型配置"
        ],
        "summary": "获取模型配置列表",
        "operationId": "get_config_list_api_config_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "页码：默认为1，必须为≥1的整数。示例：?page=2",
              "default": 1,
              "title": "Page"
            },
            "description": "页码：默认为1，必须为≥1的整数。示例：?page=2"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 100,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "每页数量：默认为10，范围1-100。示例：?page_size=20",
              "default": 10,
              "title": "Page Size"
            },
            "description": "每页数量：默认为10，范围1-100。示例：?page_size=20"
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "\n        排序规则：格式为「字段名+符号」（无符号=升序，-=降序），多字段用逗号分隔。\n        示例：?sort=create_time,-id → 按create_time升序、id降序排列\n        ",
              "title": "Sort"
            },
            "description": "\n        排序规则：格式为「字段名+符号」（无符号=升序，-=降序），多字段用逗号分隔。\n        示例：?sort=create_time,-id → 按create_time升序、id降序排列\n        "
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "\n        1. 搜索：直接传关键词（会对预设的可搜索字段进行模糊匹配）\n           示例：?search=张三 → 搜索所有含\"张三\"的记录\n           \n        2. 过滤：按「字段__操作符=值」格式传参，支持多条件（用&连接）\n           常用操作符：\n           - __contains: 包含（字符串模糊匹配）\n           - __lte: 小于等于（数字/日期）\n           - __gte: 大于等于（数字/日期）\n           - __gt: 大于（数字/日期）\n           - __lt: 小于（数字/日期）\n           \n           示例：\n           ?search=张三&age__gte=18 → 姓名含\"张三\"且年龄≥18\n           ?create_time__gte=2023-01-01 → 创建时间在2023-01-01之后\n        ",
              "title": "Search"
            },
            "description": "\n        1. 搜索：直接传关键词（会对预设的可搜索字段进行模糊匹配）\n           示例：?search=张三 → 搜索所有含\"张三\"的记录\n           \n        2. 过滤：按「字段__操作符=值」格式传参，支持多条件（用&连接）\n           常用操作符：\n           - __contains: 包含（字符串模糊匹配）\n           - __lte: 小于等于（数字/日期）\n           - __gte: 大于等于（数字/日期）\n           - __gt: 大于（数字/日期）\n           - __lt: 小于（数字/日期）\n           \n           示例：\n           ?search=张三&age__gte=18 → 姓名含\"张三\"且年龄≥18\n           ?create_time__gte=2023-01-01 → 创建时间在2023-01-01之后\n        "
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseSchema_PaginationResponse_AiModelConfigOut__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/config/{config_id}": {
      "get": {
        "tags": [
          "模型配置"
        ],
        "summary": "获取模型配置详情",
        "operationId": "get_config_api_config__config_id__get",
        "parameters": [
          {
            "name": "config_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Config Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseSchema_AiModelConfigOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "模型配置"
        ],
        "summary": "全量更新模型配置",
        "operationId": "update_config_api_config__config_id__put",
        "parameters": [
          {
            "name": "config_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Config Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AiModelConfigUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseSchema_AiModelConfigOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "模型配置"
        ],
        "summary": "局部更新模型配置",
        "operationId": "patch_config_api_config__config_id__patch",
        "parameters": [
          {
            "name": "config_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Config Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AiModelConfigPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseSchema_AiModelConfigOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "模型配置"
        ],
        "summary": "删除模型配置",
        "operationId": "delete_config_api_config__config_id__delete",
        "parameters": [
          {
            "name": "config_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Config Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/config/{config_id}/activate": {
      "post": {
        "tags": [
          "模型配置"
        ],
        "summary": "启用模型配置",
        "operationId": "activate_config_api_config__config_id__activate_post",
        "parameters": [
          {
            "name": "config_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Config Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseSchema_AiModelConfigOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AiModelConfigCreate": {
        "properties": {
          "task_type": {
            "type": "integer",
            "title": "Task Type",
            "description": "1=提取任务, 2=生成参考图, 3=生成分镜, 4=生成视频"
          },
          "name": {
            "type": "string",
            "maxLength": 100,
            "title": "Name",
            "description": "配置名称"
          },
          "base_url": {
            "type": "string",
            "maxLength": 500,
            "title": "Base Url",
            "description": "API 地址"
          },
          "api_key": {
            "type": "string",
            "maxLength": 500,
            "title": "Api Key",
            "description": "API Key"
          },
          "model": {
            "type": "string",
            "maxLength": 200,
            "title": "Model",
            "description": "模型名称"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active",
            "description": "是否启用"
          },
          "concurrency": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Concurrency",
            "description": "并发数"
          }
        },
        "type": "object",
        "required": [
          "task_type",
          "name",
          "base_url",
          "api_key",
          "model"
        ],
        "title": "AiModelConfigCreate",
        "description": "创建请求：必填字段。"
      },
      "AiModelConfigOut": {
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "创建时间"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "更新时间"
          },
          "task_type": {
            "type": "integer",
            "title": "Task Type",
            "description": "1=提取任务, 2=生成参考图, 3=生成分镜, 4=生成视频"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "配置名称"
          },
          "base_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Url",
            "description": "API 地址"
          },
          "api_key": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key",
            "description": "API Key"
          },
          "model": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Model",
            "description": "模型名称"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "description": "是否启用"
          },
          "concurrency": {
            "type": "integer",
            "title": "Concurrency",
            "description": "并发数"
          },
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "配置ID"
          }
        },
        "type": "object",
        "required": [
          "created_at",
          "updated_at",
          "task_type",
          "name",
          "is_active",
          "concurrency",
          "id"
        ],
        "title": "AiModelConfigOut",
        "description": "配置输出。"
      },
      "AiModelConfigPatch": {
        "properties": {
          "task_type": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Type",
            "description": "1=提取任务, 2=生成参考图, 3=生成分镜, 4=生成视频"
          },
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Name",
            "description": "配置名称"
          },
          "base_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Url",
            "description": "API 地址"
          },
          "api_key": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key",
            "description": "API Key"
          },
          "model": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Model",
            "description": "模型名称"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active",
            "description": "是否启用"
          },
          "concurrency": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Concurrency",
            "description": "并发数"
          }
        },
        "type": "object",
        "title": "AiModelConfigPatch",
        "description": "局部更新：全字段可选。"
      },
      "AiModelConfigUpdate": {
        "properties": {
          "task_type": {
            "type": "integer",
            "title": "Task Type",
            "description": "1=提取任务, 2=生成参考图, 3=生成分镜, 4=生成视频"
          },
          "name": {
            "type": "string",
            "maxLength": 100,
            "title": "Name",
            "description": "配置名称"
          },
          "base_url": {
            "type": "string",
            "maxLength": 500,
            "title": "Base Url",
            "description": "API 地址"
          },
          "api_key": {
            "type": "string",
            "maxLength": 500,
            "title": "Api Key",
            "description": "API Key"
          },
          "model": {
            "type": "string",
            "maxLength": 200,
            "title": "Model",
            "description": "模型名称"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active",
            "description": "是否启用"
          },
          "concurrency": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Concurrency",
            "description": "并发数"
          }
        },
        "type": "object",
        "required": [
          "task_type",
          "name",
          "base_url",
          "api_key",
          "model"
        ],
        "title": "AiModelConfigUpdate",
        "description": "全量更新：同创建。"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Pagination": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "pages": {
            "type": "integer",
            "title": "Pages"
          }
        },
        "type": "object",
        "required": [
          "total",
          "page",
          "page_size",
          "pages"
        ],
        "title": "Pagination"
      },
      "PaginationResponse_AiModelConfigOut_": {
        "properties": {
          "items": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/AiModelConfigOut"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Items"
          },
          "pagination": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Pagination"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "PaginationResponse[AiModelConfigOut]"
      },
...
      "ResponseSchema": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 0
          },
          "data": {
            "anyOf": [
              {

              },
              {
                "items": {

                },
                "type": "array"
              },
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "操作成功"
          }
        },
        "type": "object",
        "title": "ResponseSchema",
        "description": "统一API响应格式\n\n属性:\n    code (int): 响应状态码，0表示成功，非0表示异常\n    data (T): 响应数据，泛型类型\n    message (str): 响应消息",
        "example": {
          "code": 0,
          "data": {

          },
          "message": "操作成功"
        }
      },
      "ResponseSchema_AiModelConfigOut_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 0
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AiModelConfigOut"
              },
              {
                "items": {

                },
                "type": "array"
              },
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "操作成功"
          }
        },
        "type": "object",
        "title": "ResponseSchema[AiModelConfigOut]",
        "example": {
          "code": 0,
          "data": {

          },
          "message": "操作成功"
        }
      },
...
      "ResponseSchema_PaginationResponse_AiModelConfigOut__": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 0
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PaginationResponse_AiModelConfigOut_"
              },
              {
                "items": {

                },
                "type": "array"
              },
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "操作成功"
          }
        },
        "type": "object",
        "title": "ResponseSchema[PaginationResponse[AiModelConfigOut]]",
        "example": {
          "code": 0,
          "data": {

          },
          "message": "操作成功"
        }
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}